+++ packages/datastructure/src/queue/queue.ts
@@ -1,3 +1,11 @@
+class QueueNode<T> {
+  data: T;
+  next: QueueNode<T> | undefined;
+  constructor(data: T) {
+    this.data = data;
+  }
+}
+
 /**
  * @license MIT
  * @copyright 2024 Thinh Trinh Duc <thinh.duc.trinh@gmail.com>
@@ -5,13 +13,26 @@
  * @class
  */
 export class Queue<T> {
-  #array: T[] = [];
+  #root: QueueNode<T> | undefined;
+  #top: QueueNode<T> | undefined;
+  #count: number = 0;
 
   /**
    * Create a new queue
    */
   constructor(...elements: T[]) {
-    this.#array = [...elements];
+    this.enqueue(...elements);
+  }
+
+  #enqueue(data: T) {
+    if (!this.#top) {
+      this.#root = this.#top = new QueueNode(data);
+    } else {
+      this.#top.next = new QueueNode(data);
+      this.#top = this.#top.next;
+    }
+
+    this.#count++;
   }
 
   /**
@@ -26,32 +47,27 @@ export class Queue<T> {
   }
 
   /**
-   * Convert stack to array
-   */
-  toArray() {
-    return [...this.#array];
-  }
-
-  /**
    * Queue's size
    * @readonly
    */
   get size() {
-    return this.#array.length;
+    return this.#count;
   }
 
   /**
    * Check if the queue is empty.
    */
   isEmpty() {
-    return this.#array.length === 0;
+    return this.#root === undefined;
   }
 
   /**
    * Enqueue new elements, and returns the new size of the queue.
    */
-  enqueue(...value: T[]) {
-    return this.#array.push(...value);
+  enqueue(...values: T[]) {
+    for (const value of values) {
+      this.#enqueue(value);
+    }
   }
 
   /**
@@ -60,21 +76,50 @@ export class Queue<T> {
    * If the queue is empty, undefined is returned and the queue is not modified.
    */
   dequeue() {
-    return this.#array.shift();
+    if (!this.#root) return undefined;
+
+    const value = this.#root.data;
+    this.#root = this.#root.next;
+
+    if (this.#root === undefined) {
+      this.#top = undefined;
+    }
+
+    this.#count--;
+
+    return value;
   }
 
   /**
    * Returns the top element without removing it
    */
   peek() {
-    return this.#array[0];
+    return this.#root?.data;
   }
 
   /**
    * Clear queue
    */
   clear() {
-    this.#array = [];
+    this.#root = undefined;
+    this.#top = undefined;
+    this.#count = 0;
     return this;
+  }
+
+  /**
+   * Convert stack to array
+   */
+  toArray() {
+    return [...this];
+  }
+
+  *[Symbol.iterator]() {
+    let curr = this.#root;
+
+    while (curr) {
+      yield curr.data;
+      curr = curr.next;
+    }
   }
 }