+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # docs
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @ultify/datastructure@0.4.2
+
 ## 0.0.10
 
 ### Patch Changes